requirements.txt
streamlit
tika
langchain
langchain_huggingface
langchain_community
langchain_groq
python-dotenv
chromadb==0.5.3
spacy
requests
rank_bm25
psycopg2
langsmith
streamlit_extras

supabase
psycopg2
streamlit_authenticator
streamlit_extras
nltk


# For downloading the model
# python -m spacy download en_core_web_sm


# For downloading the NLTK resources
# run "setup_nltk.py"